Factor utility methods out of the main registry module
